+++ app/src/main/java/pl/olszak/todo/addition/model/AddTaskIntent.kt
@@ -1,0 +1,8 @@
+package pl.olszak.todo.addition.model
+
+import pl.olszak.todo.core.Intention
+
+sealed class AddTaskIntent : Intention {
+
+    data class ProcessTask(val taskTitle: String) : AddTaskIntent()
+}
