+++ app/src/main/java/pl/olszak/todo/addition/model/AddTaskResult.kt
@@ -1,0 +1,12 @@
+package pl.olszak.todo.addition.model
+
+import pl.olszak.todo.core.Result
+
+sealed class AddTaskResult : Result {
+
+    object Pending : AddTaskResult()
+
+    object Added : AddTaskResult()
+
+    object Failure : AddTaskResult()
+}
