--- app/src/main/java/pl/olszak/todo/feature/todos/adapter/TasksDelegate.kt
@@ -1,23 +1,0 @@
-package pl.olszak.todo.feature.todos.adapter
-
-import android.widget.TextView
-import com.hannesdorfmann.adapterdelegates4.dsl.adapterDelegateViewBinding
-import pl.olszak.todo.R
-import pl.olszak.todo.core.adapter.AdapterItem
-import pl.olszak.todo.databinding.ViewHolderTaskBinding
-import pl.olszak.todo.feature.todos.model.TaskViewItem
-
-fun createTaskDelegate() =
-    adapterDelegateViewBinding<TaskViewItem, AdapterItem, ViewHolderTaskBinding>(
-        { layoutInflater, root ->
-            ViewHolderTaskBinding.inflate(layoutInflater, root, false)
-        }
-    ) {
-        val title: TextView = itemView.findViewById(R.id.title)
-        val description: TextView = itemView.findViewById(R.id.description)
-
-        bind {
-            title.text = item.title
-            description.text = item.description
-        }
-    }
