--- app/src/main/java/pl/olszak/todo/feature/todos/model/TodosViewState.kt
@@ -1,7 +1,0 @@
-package pl.olszak.todo.feature.todos.model
-
-import pl.olszak.todo.core.ViewState
-
-data class TodosViewState(
-    val tasks: List<TaskViewItem> = emptyList()
-) : ViewState
