+++ app/src/main/java/pl/olszak/todo/todos/model/TodosViewState.kt
@@ -1,0 +1,7 @@
+package pl.olszak.todo.todos.model
+
+import pl.olszak.todo.core.ViewState
+
+data class TodosViewState(
+    val tasks: List<TaskViewItem> = emptyList()
+) : ViewState
