--- app/src/main/res/drawable/circle_shape.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-    <size
-        android:width="@dimen/icon_size"
-        android:height="@dimen/icon_size" />
-    <solid android:color="?attr/colorPrimary" />
-</shape>
